fix more issues with known types

+++ db4oPad/UserAssembliesProvider.cs
@@ -63,10 +63,40 @@ namespace Gamlor.Db4oPad
             foreach (var path in paths)
             {
                 var targetLocation = targetDir + @"\" + Path.GetFileName(path);
-                File.Copy(path, targetLocation,true);
+                if (!File.Exists(targetLocation) || !IsFileLocked(new FileInfo(targetLocation)))
+                {
+                    File.Copy(path, targetLocation, true);
+                }
                 result.Add(targetLocation);
             }
             return result;
+        }
+
+        private static bool IsFileLocked(FileInfo file)
+        {
+            
+            FileStream stream = null;
+
+            try
+            {
+                stream = file.Open(FileMode.Open, FileAccess.ReadWrite, FileShare.None);
+            }
+            catch (IOException)
+            {
+                //the file is unavailable because it is:
+                //still being written to
+                //or being processed by another thread
+                //or does not exist (has already been processed)
+                return true;
+            }
+            finally
+            {
+                if (stream != null)
+                    stream.Close();
+            }
+
+            //file is not locked
+            return false;
         }
 
         private static string GetContextPath()
